refactor(review): address comments
