+++ src/directives/format-length.directive.ts
@@ -1,4 +1,4 @@
-import i18next from 'i18next'
+import { formatLength } from '@/services/common/formatting.utils'
 import { App } from 'vue'
 
 export default {
@@ -18,17 +18,4 @@ function format(el: HTMLElement, value: number): void {
   let formattedText: string = ''
   formattedText = formatLength(value)
   el.textContent = formattedText
-}
-
-export function formatLength(value: number): string {
-  //null covers API errors or unaivalble data (eg. elevation)
-  if (value === null) {
-    return i18next.t('N/A', { ns: 'client' })
-  } else if (value < 1000) {
-    return `${value.toFixed(2)} m`
-  } else if (value >= 1000) {
-    return `${(value / 1000).toFixed(2)} km`
-  } else {
-    return ''
-  }
 }
